[Fix] coupons : Integer에서 가격 자료형 Double로 변경 #13

+++ coupon/src/main/java/com/haot/coupon/application/dto/request/coupons/CouponCreateRequest.java
@@ -11,10 +11,10 @@ public record CouponCreateRequest(
         LocalDateTime couponExpiredDate,
         CouponType couponType,
         DiscountPolicy discountPolicy,
-        Integer minimumAmount,
-        Integer maximumAmount,
-        Integer discountRate,
-        Integer discountAmount,
+        Double minimumAmount,
+        Double maximumAmount,
+        Double discountRate,
+        Double discountAmount,
         Integer maxQuantity,
         Integer issuedQuantity
 ) {
