+++ coupon/src/main/java/com/haot/coupon/application/dto/response/coupons/ReservationVerifyResponse.java
@@ -6,6 +6,6 @@ import lombok.Builder;
 public record ReservationVerifyResponse(
 
         String reservationCouponId,
-        Integer discountedPrice
+        Double discountedPrice
 ) {
 }
